Rename persons.py to people.py; remove add_person function
