+++ models/people.py
@@ -1,0 +1,17 @@
+import random
+from rooms import Room, Office, Living
+
+class Person(object):
+    """
+    Creates a Person object. Classes Fellow and Staff inherit from it.
+    """
+
+class Staff(Person):
+
+    job_type = "Staff"
+
+    ''' Allocation of Staff to Office '''
+
+class Fellow(Person):
+
+    job_type = "Fellow"
